Add stack module pointer
The module pointer refers to the current module during execution and is preserved on the stack during function calls. This allows for memory and table instructions to be shorter as they do not need absolute pointers anymore. The entry instruction is responsible for setting the local module pointer, which now needs to be emitted if either locals or any instruction requiring the module pointer is present.