Be strict and explicit about different indices
WASM defines a bunch of different kinds of indices, plus the implementation of the linker also adds additional ones. To better keep track where which kind of index is used, now incompatible types are used. Conversions from one kind to another need to be done explicitly to prevent accidental ones.